Refactored the collective pallet
